+++ core/src/tests/common/helpers.rs
@@ -41,6 +41,7 @@ use std::sync::Arc;
 use kvdb::{MockDbRepository, DBTransaction, KeyValueDB, MemoryDBRepository};
 use db;
 
+
 // TODO: move everything over to get_null_spec.
 pub fn get_test_spec() -> Spec { Spec::new_test() }
