+++ core/src/tests/common/mod.rs
@@ -1,0 +1,4 @@
+pub mod helpers;
+pub mod test_client;
+
+pub use self::test_client::{EachBlockWith, TestBlockChainClient};
