+++ core/src/tests/mod.rs
@@ -22,5 +22,5 @@
 
 mod client;
 mod executive;
-pub mod helpers;
+pub mod common;
