+++ core/tests/src/lib.rs
@@ -1,0 +1,13 @@
+#[cfg(test)]
+extern crate acore;
+#[cfg(test)]
+extern crate aion_types;
+#[cfg(test)]
+extern crate key;
+#[cfg(test)]
+extern crate db as kvdb;
+#[cfg(test)]
+extern crate acore_io as io;
+#[cfg(test)]
+extern crate tempdir;
+
